+++ src/main/java/com/bod/bod/admin/controller/AdminController.java
@@ -6,7 +6,6 @@ import com.bod.bod.admin.dto.AdminChallengeResponseDto;
 import com.bod.bod.admin.dto.AdminChallengeUpdateRequestDto;
 import com.bod.bod.admin.dto.AdminChallengeUpdateResponseDto;
 import com.bod.bod.admin.dto.AdminChallengesResponseDto;
-import com.bod.bod.admin.dto.AdminPaginationResponseDto;
 import com.bod.bod.admin.dto.AdminUserStatusUpdateRequestDto;
 import com.bod.bod.admin.dto.AdminUserStatusUpdateResponseDto;
 import com.bod.bod.admin.dto.AdminUserUpdateRequestDto;
@@ -14,14 +13,11 @@ import com.bod.bod.admin.dto.AdminUserUpdateResponseDto;
 import com.bod.bod.admin.dto.AdminUsersResponseDto;
 import com.bod.bod.admin.dto.AdminVerificationGetResponse;
 import com.bod.bod.admin.service.AdminService;
-import com.bod.bod.challenge.entity.Challenge;
 import com.bod.bod.global.dto.CommonResponseDto;
+import com.bod.bod.global.dto.PaginationResponse;
 import com.bod.bod.global.jwt.security.UserDetailsImpl;
-import com.bod.bod.verification.entity.Verification;
 import jakarta.validation.Valid;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -46,12 +42,12 @@ public class AdminController {
     private final AdminService adminService;
 
     @GetMapping("/admins/users")
-    public ResponseEntity<CommonResponseDto<AdminPaginationResponseDto<AdminUsersResponseDto>>> getAllUsers(
+    public ResponseEntity<CommonResponseDto<PaginationResponse<AdminUsersResponseDto>>> getAllUsers(
         @RequestParam(value = "page", defaultValue = "1") int page,
         @RequestParam(value = "size", defaultValue = "10") int size,
         @AuthenticationPrincipal UserDetailsImpl userDetails
     ) {
-        AdminPaginationResponseDto<AdminUsersResponseDto> usersList = adminService.getAllUsers(page-1,size,userDetails.getUser());
+        PaginationResponse<AdminUsersResponseDto> usersList = adminService.getAllUsers(page-1,size,userDetails.getUser());
         return ResponseEntity.ok().body(new CommonResponseDto<>
             (HttpStatus.OK.value(), "전체 회원 조회에 성공하였습니다!", usersList));
     }
@@ -111,13 +107,13 @@ public class AdminController {
     }
 
     @GetMapping("/admins/challenges/{challengeId}/verifications")
-    public ResponseEntity<CommonResponseDto<AdminPaginationResponseDto<AdminVerificationGetResponse>>> getVerifications(
+    public ResponseEntity<CommonResponseDto<PaginationResponse<AdminVerificationGetResponse>>> getVerifications(
         @PathVariable(name = "challengeId") Long challengeId,
         @RequestParam(value = "page", defaultValue = "1") int page,
         @RequestParam(value = "size", defaultValue = "10") int size,
         @AuthenticationPrincipal UserDetailsImpl userDetails
     ) {
-        AdminPaginationResponseDto<AdminVerificationGetResponse> verifications = adminService.getVerifications(challengeId, page - 1, size, userDetails.getUser());
+        PaginationResponse<AdminVerificationGetResponse> verifications = adminService.getVerifications(challengeId, page - 1, size, userDetails.getUser());
 
         return ResponseEntity.ok().body(new CommonResponseDto<>
             (HttpStatus.OK.value(), "챌린지 인증 요청 목록 조회에 성공하였습니다!", verifications));
@@ -144,12 +140,12 @@ public class AdminController {
     }
 
     @GetMapping("/admins/challenges")
-    public ResponseEntity<CommonResponseDto<AdminPaginationResponseDto<AdminChallengesResponseDto>>> getAllChallenges(
+    public ResponseEntity<CommonResponseDto<PaginationResponse<AdminChallengesResponseDto>>> getAllChallenges(
         @RequestParam(value = "page", defaultValue = "1") int page,
         @RequestParam(value = "size", defaultValue = "10") int size,
         @AuthenticationPrincipal UserDetailsImpl userDetails
     ) {
-        AdminPaginationResponseDto<AdminChallengesResponseDto> challengeList = adminService.getAllChallenges(page - 1, size, userDetails.getUser());
+        PaginationResponse<AdminChallengesResponseDto> challengeList = adminService.getAllChallenges(page - 1, size, userDetails.getUser());
 
         return ResponseEntity.ok().body(new CommonResponseDto<>
             (HttpStatus.OK.value(), "전체 챌린지 조회에 성공하였습니다!", challengeList));
