--- src/main/java/com/bod/bod/admin/dto/AdminPaginationResponseDto.java
@@ -1,23 +1,0 @@
-package com.bod.bod.admin.dto;
-
-import java.util.List;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class AdminPaginationResponseDto<T> {
-    private List<T> content;
-    private int totalPages;
-    private long totalElements;
-    private int currentPage;
-    private int pageSize;
-
-    public AdminPaginationResponseDto(List<T> content, int totalPages, long totalElements, int currentPage, int pageSize) {
-        this.content = content;
-        this.totalPages = totalPages;
-        this.totalElements = totalElements;
-        this.currentPage = currentPage;
-        this.pageSize = pageSize;
-    }
-}
