+++ src/main/java/com/bod/bod/admin/service/AdminService.java
@@ -9,7 +9,6 @@ import com.bod.bod.admin.dto.AdminChallengeResponseDto;
 import com.bod.bod.admin.dto.AdminChallengeUpdateRequestDto;
 import com.bod.bod.admin.dto.AdminChallengeUpdateResponseDto;
 import com.bod.bod.admin.dto.AdminChallengesResponseDto;
-import com.bod.bod.admin.dto.AdminPaginationResponseDto;
 import com.bod.bod.admin.dto.AdminUserStatusUpdateRequestDto;
 import com.bod.bod.admin.dto.AdminUserStatusUpdateResponseDto;
 import com.bod.bod.admin.dto.AdminUserUpdateRequestDto;
@@ -20,6 +19,7 @@ import com.bod.bod.challenge.entity.Category;
 import com.bod.bod.challenge.entity.Challenge;
 import com.bod.bod.challenge.entity.ConditionStatus;
 import com.bod.bod.challenge.repository.ChallengeRepository;
+import com.bod.bod.global.dto.PaginationResponse;
 import com.bod.bod.global.exception.ErrorCode;
 import com.bod.bod.global.exception.FileUploadFailureException;
 import com.bod.bod.global.exception.GlobalException;
@@ -29,14 +29,12 @@ import com.bod.bod.user.repository.UserRepository;
 import com.bod.bod.verification.entity.Verification;
 import com.bod.bod.verification.repository.VerificationRepository;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -56,7 +54,7 @@ public class AdminService {
     @Value("${cloud.aws.s3.bucket}")
     private String BUCKET;
 
-    public AdminPaginationResponseDto<AdminUsersResponseDto> getAllUsers(int page, int size, User loginUser) {
+    public PaginationResponse<AdminUsersResponseDto> getAllUsers(int page, int size, User loginUser) {
         if (!loginUser.getUserRole().equals(UserRole.ADMIN)) {
             throw new GlobalException(ErrorCode.USER_ACCESS_DENIED);
         }
@@ -66,7 +64,7 @@ public class AdminService {
 
         List<AdminUsersResponseDto> usersDto = users.getContent().stream().map(AdminUsersResponseDto::new).toList();
 
-        return new AdminPaginationResponseDto<>(
+        return new PaginationResponse<>(
             usersDto,
             users.getTotalPages(),
             users.getTotalElements(),
@@ -158,7 +156,7 @@ public class AdminService {
         challengeRepository.delete(challenge);
     }
 
-    public AdminPaginationResponseDto<AdminVerificationGetResponse> getVerifications(long challengeId, int page, int size, User loginUser) {
+    public PaginationResponse<AdminVerificationGetResponse> getVerifications(long challengeId, int page, int size, User loginUser) {
         if (!loginUser.getUserRole().equals(UserRole.ADMIN)) {
             throw new GlobalException(ErrorCode.USER_ACCESS_DENIED);
         }
@@ -170,7 +168,7 @@ public class AdminService {
         }
         List<AdminVerificationGetResponse> verificationResponses = verifications.getContent().stream().map(AdminVerificationGetResponse::new).toList();
 
-        return new AdminPaginationResponseDto<>(
+        return new PaginationResponse<>(
             verificationResponses,
             verifications.getTotalPages(),
             verifications.getTotalElements(),
@@ -213,7 +211,7 @@ public class AdminService {
         }
     }
 
-    public AdminPaginationResponseDto<AdminChallengesResponseDto> getAllChallenges(int page, int size, User loginUser) {
+    public PaginationResponse<AdminChallengesResponseDto> getAllChallenges(int page, int size, User loginUser) {
         if (!loginUser.getUserRole().equals(UserRole.ADMIN)) {
             throw new GlobalException(ErrorCode.USER_ACCESS_DENIED);
         }
@@ -221,7 +219,7 @@ public class AdminService {
         Page<Challenge> challenges = challengeRepository.findAllByOrderByCreatedAtAsc(pageable);
         List<AdminChallengesResponseDto> challengesResponseDtos = challenges.getContent().stream().map(AdminChallengesResponseDto::new).toList();
 
-        return new AdminPaginationResponseDto<>(
+        return new PaginationResponse<>(
             challengesResponseDtos,
             challenges.getTotalPages(),
             challenges.getTotalElements(),
