+++ src/main/java/com/bod/bod/challenge/controller/ChallengeController.java
@@ -3,10 +3,10 @@ package com.bod.bod.challenge.controller;
 import com.bod.bod.challenge.dto.ChallengeResponseDto;
 import com.bod.bod.challenge.dto.ChallengeSummaryResponseDto;
 import com.bod.bod.challenge.dto.ChallengeUserListDto;
-import com.bod.bod.challenge.dto.PaginationResponse;
 import com.bod.bod.challenge.entity.Category;
 import com.bod.bod.challenge.service.ChallengeService;
 import com.bod.bod.global.dto.CommonResponseDto;
+import com.bod.bod.global.dto.PaginationResponse;
 import com.bod.bod.global.jwt.security.UserDetailsImpl;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
